+++ .gitignore
@@ -65,3 +65,7 @@ src/UniGetUI/choco-cli/redirects/cup.exe.ignore
 src/UniGetUI/choco-cli/bin/ssh-copy-id.exe
 src/UniGetUI/choco-cli/extensions/chocolatey-visualstudio/
 src/UniGetUI/choco-cli/extensions/chocolatey-dotnetfx/
+
+# Ignore compiled binaries
+**/bin/
+**/obj/
