+++ src/UniGetUI/PackageEngine/Classes/BundleFormatType.cs
@@ -1,0 +1,10 @@
+﻿namespace UniGetUI.PackageEngine.Classes
+{
+    public enum BundleFormatType
+    {
+        JSON,
+        YAML,
+        XML
+    }
+
+}
