+++ src/UniGetUI/PackageEngine/Classes/DeserializedPackageStatus.cs
@@ -1,0 +1,12 @@
+﻿namespace UniGetUI.PackageEngine.Classes
+{
+    public enum DeserializedPackageStatus
+    {
+        ManagerNotFound,
+        ManagerNotEnabled,
+        ManagerNotReady,
+        SourceNotFound,
+        IsAvailable
+    }
+
+}
