+++ src/UniGetUI/PackageEngine/Classes/ManagerCapabilities.cs
@@ -1,0 +1,24 @@
+﻿using System.Runtime.InteropServices;
+
+namespace UniGetUI.PackageEngine.Classes
+{
+    public struct ManagerCapabilities
+    {
+        public bool CanRunAsAdmin = false;
+        public bool CanSkipIntegrityChecks = false;
+        public bool CanRunInteractively = false;
+        public bool CanRemoveDataOnUninstall = false;
+        public bool SupportsCustomVersions = false;
+        public bool SupportsCustomArchitectures = false;
+        public Architecture[] SupportedCustomArchitectures = new Architecture[0];
+        public bool SupportsCustomScopes = false;
+        public bool SupportsPreRelease = false;
+        public bool SupportsCustomLocations = false;
+        public bool SupportsCustomSources = false;
+        public ManagerSource.Capabilities Sources { get; set; }
+        public ManagerCapabilities()
+        {
+            Sources = new ManagerSource.Capabilities();
+        }
+    }
+}
