+++ src/UniGetUI/PackageEngine/Classes/ManagerProperties.cs
@@ -1,0 +1,16 @@
+﻿namespace UniGetUI.PackageEngine.Classes
+{
+    public class ManagerProperties
+    {
+        public string Name { get; set; } = "Unset";
+        public string Description { get; set; } = "Unset";
+        public string IconId { get; set; } = "Unset";
+        public string ColorIconId { get; set; } = "Unset";
+        public string ExecutableCallArgs { get; set; } = "Unset";
+        public string ExecutableFriendlyName { get; set; } = "Unset";
+        public string InstallVerb { get; set; } = "Unset";
+        public string UpdateVerb { get; set; } = "Unset";
+        public string UninstallVerb { get; set; } = "Unset";
+
+    }
+}
