+++ src/UniGetUI/PackageEngine/Classes/ManagerStatus.cs
@@ -1,0 +1,11 @@
+﻿namespace UniGetUI.PackageEngine.Classes
+{
+    public class ManagerStatus
+    {
+        public string Version = "";
+        public bool Found = false;
+        public string ExecutablePath = "";
+        public ManagerStatus()
+        { }
+    }
+}
