+++ src/UniGetUI/PackageEngine/Classes/Package.cs
@@ -1,42 +1,14 @@
-﻿using UniGetUI.Core.Data;
-using UniGetUI.Core;
-using System;
-using System.Collections.Generic;
+﻿using System;
 using System.ComponentModel;
 using System.IO;
 using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-using System.Text.Json;
 using System.Text.Json.Nodes;
 using System.Threading.Tasks;
-using Windows.ApplicationModel;
-using Windows.Devices.Bluetooth.Advertisement;
-using System.ComponentModel.Design;
+using UniGetUI.Core;
+using UniGetUI.Core.Data;
 
 namespace UniGetUI.PackageEngine.Classes
 {
-    /// <summary>
-    /// Represents the scope of a package. To be coherent with package manager naming, the values are repeated.
-    /// </summary>
-    public enum PackageScope
-    {
-        // Repeated entries for coherence with Package Managers
-        Global = 1,
-        Machine = 1,
-        Local = 0,
-        User = 0,
-    }
-
-    public enum PackageTag
-    {
-        Default,
-        AlreadyInstalled,
-        IsUpgradable,
-        Pinned,
-        OnQueue,
-        BeingProcessed,
-        Failed
-    }
 
     /// <summary>
     /// This class represents a installable package or a package that is already installed.
@@ -427,354 +399,5 @@ namespace UniGetUI.PackageEngine.Classes
             Tag = tag;
         }
 
-    }
-
-    public class UpgradablePackage : Package
-    {
-        // Public properties
-        public float NewVersionAsFloat { get; }
-        public override bool IsUpgradable { get; } = true;
-
-        private string __hash;
-
-        /// <summary>
-        /// Creates an UpgradablePackage object representing a package that can be upgraded; given its name, id, installed version, new version, source and manager, and an optional scope.
-        /// </summary>
-        /// <param name="name"></param>
-        /// <param name="id"></param>
-        /// <param name="installed_version"></param>
-        /// <param name="new_version"></param>
-        /// <param name="source"></param>
-        /// <param name="manager"></param>
-        /// <param name="scope"></param>
-        public UpgradablePackage(string name, string id, string installed_version, string new_version, ManagerSource source, PackageManager manager, PackageScope scope = PackageScope.Local) : base(name, id, installed_version, source, manager, scope)
-        {
-            NewVersion = new_version;
-            IsChecked = true;
-            NewVersionAsFloat = GetFloatNewVersion();
-
-            __hash = Manager.Name + "\\" + Source.Name + "\\" + Id + "\\" + Version + "->" + NewVersion;
-        }
-        public new string GetHash()
-        {
-            return __hash;
-        }
-
-        /// <summary>
-        /// Returns a float value representing the new new version of the package, for comparison purposes.
-        /// </summary>
-        /// <returns></returns>
-        public float GetFloatNewVersion()
-        {
-            string _ver = "";
-            bool _dotAdded = false;
-            foreach (char _char in NewVersion)
-            {
-                if (char.IsDigit(_char))
-                    _ver += _char;
-                else if (_char == '.')
-                {
-                    if (!_dotAdded)
-                    {
-                        _ver += _char;
-                        _dotAdded = true;
-                    }
-                }
-            }
-            float res = 0.0F;
-            if (_ver != "" && _ver != ".")
-                try
-                {
-                    return float.Parse(_ver);
-                }
-                catch (Exception)
-                {
-                }
-            return res;
-        }
-
-        /// <summary>
-        /// This version will check if the new version of the package is already present 
-        /// on the InstalledPackages list, to prevent already installed updates from being updated again.
-        /// </summary>
-        /// <returns></returns>
-        public bool NewVersionIsInstalled()
-        {
-            foreach (Package package in Tools.App.MainWindow.NavigationPage.InstalledPage.Packages)
-                if (package.Manager == Manager && package.Id == Id && package.Version == NewVersion && package.Source.Name == Source.Name)
-                    return true;
-            return false;
-        }
-    }
-
-    /// <summary>
-    /// The properties of a given package.
-    /// </summary>
-    public class PackageDetails
-    {
-        public Package Package { get; }
-        public string Name { get; }
-        public string Id { get; }
-        public string Version { get; }
-        public string NewVersion { get; }
-        public ManagerSource Source { get; }
-        public string Description { get; set; } = "";
-        public string Publisher { get; set; } = "";
-        public string Author { get; set; } = "";
-        public Uri HomepageUrl { get; set; } = null;
-        public string License { get; set; } = "";
-        public Uri LicenseUrl { get; set; } = null;
-        public Uri InstallerUrl { get; set; } = null;
-        public string InstallerHash { get; set; } = "";
-        public string InstallerType { get; set; } = "";
-        public double InstallerSize { get; set; } = 0; // In Megabytes
-        public Uri ManifestUrl { get; set; } = null;
-        public string UpdateDate { get; set; } = "";
-        public string ReleaseNotes { get; set; } = "";
-        public Uri ReleaseNotesUrl { get; set; } = null;
-        public string[] Tags { get; set; } = new string[0];
-
-        /// <summary>
-        /// Construct a PackageDetails object from a given package. The constructor does 
-        /// NOT load the package's details. They must be loaded manually
-        /// </summary>
-        /// <param name="package"></param>
-        public PackageDetails(Package package)
-        {
-            Package = package;
-            Name = package.Name;
-            Id = package.Id;
-            Version = package.Version;
-            Source = package.Source;
-            if (package is UpgradablePackage)
-                NewVersion = ((UpgradablePackage)package).NewVersion;
-            else
-                NewVersion = "";
-        }
-    }
-
-    /// <summary>
-    /// This class represents the options in which a package must be installed, updated or uninstalled.
-    /// </summary>
-    public class InstallationOptions
-    {
-        public bool SkipHashCheck { get; set; } = false;
-        public bool InteractiveInstallation { get; set; } = false;
-        public bool RunAsAdministrator { get; set; } = false;
-        public string Version { get; set; } = "";
-        public Architecture? Architecture { get; set; } = null;
-        public PackageScope? InstallationScope { get; set; } = null;
-        public List<string> CustomParameters { get; set; } = new List<string>();
-        public bool RemoveDataOnUninstall { get; set; } = false;
-        public bool PreRelease { get; set; } = false;
-        public string CustomInstallLocation { get; set; } = "";
-
-        public Package Package { get; }
-
-        private string _saveFileName = "Unknown.Unknown.InstallationOptions";
-
-        /// <summary>
-        /// Construct a new InstallationOptions object for a given package. The options will be 
-        /// loaded from disk unless the reset parameter is set to true, in which case the options
-        /// will be the default ones.
-        /// </summary>
-        /// <param name="package"></param>
-        /// <param name="reset"></param>
-        public InstallationOptions(Package package, bool reset = false)
-        {
-            Package = package;
-            _saveFileName = Package.Manager.Name.Replace(" ", "").Replace(".", "") + "." + Package.Id;
-            if (!reset)
-            {
-                LoadOptionsFromDisk();
-            }
-        }
-
-        /// <summary>
-        /// Returns a new InstallationOptions object from a given package. The options will be
-        /// loaded from the disk asynchronously.
-        /// </summary>
-        /// <param name="package"></param>
-        /// <returns></returns>
-        public static async Task<InstallationOptions> FromPackageAsync(Package package)
-        {
-            var options = new InstallationOptions(package, reset: true);
-            await options.LoadOptionsFromDiskAsync();
-            return options;
-        }
-
-        /// <summary>
-        /// Overload of the constructor that accepts an UpgradablePackage object.
-        /// </summary>
-        /// <param name="package"></param>
-        /// <param name="reset"></param>
-        public InstallationOptions(UpgradablePackage package, bool reset = false) : this((Package)package, reset)
-        { }
-
-        /// <summary>
-        /// Returns a new InstallationOptions object from a given SerializableInstallationOptions_v1 and a package.
-        /// </summary>
-        /// <param name="options"></param>
-        /// <param name="package"></param>
-        /// <returns></returns>
-        public static InstallationOptions FromSerialized(SerializableInstallationOptions_v1 options, Package package)
-        {
-            InstallationOptions opt = new(package, reset: true);
-            opt.FromSerialized(options);
-            return opt;
-        }
-
-        /// <summary>
-        /// Loads and applies the options from the given SerializableInstallationOptions_v1 object to the current object.
-        /// </summary>
-        /// <param name="options"></param>
-        public void FromSerialized(SerializableInstallationOptions_v1 options)
-        {
-            SkipHashCheck = options.SkipHashCheck;
-            InteractiveInstallation = options.InteractiveInstallation;
-            RunAsAdministrator = options.RunAsAdministrator;
-            CustomInstallLocation = options.CustomInstallLocation;
-            Version = options.Version;
-            PreRelease = options.PreRelease;
-            if (options.Architecture != "" && CommonTranslations.InvertedArchNames.ContainsKey(options.Architecture))
-                Architecture = CommonTranslations.InvertedArchNames[options.Architecture];
-            if (options.InstallationScope != "" && CommonTranslations.InvertedScopeNames_NonLang.ContainsKey(options.InstallationScope))
-                InstallationScope = CommonTranslations.InvertedScopeNames_NonLang[options.InstallationScope];
-            CustomParameters = options.CustomParameters;
-        }
-
-        /// <summary>
-        /// Returns a SerializableInstallationOptions_v1 object containing the options of the current instance.
-        /// </summary>
-        /// <returns></returns>
-        public SerializableInstallationOptions_v1 Serialized()
-        {
-            SerializableInstallationOptions_v1 options = new();
-            options.SkipHashCheck = SkipHashCheck;
-            options.InteractiveInstallation = InteractiveInstallation;
-            options.RunAsAdministrator = RunAsAdministrator;
-            options.CustomInstallLocation = CustomInstallLocation;
-            options.PreRelease = PreRelease;
-            options.Version = Version;
-            if (Architecture != null)
-                options.Architecture = CommonTranslations.ArchNames[Architecture.Value];
-            if (InstallationScope != null)
-                options.InstallationScope = CommonTranslations.ScopeNames_NonLang[InstallationScope.Value];
-            options.CustomParameters = CustomParameters;
-            return options;
-        }
-
-        private FileInfo GetPackageOptionsFile()
-        {
-            var optionsFileName = Package.Manager.Name + "." + Package.Id + ".json";
-            return new FileInfo(Path.Join(CoreData.UniGetUIInstallationOptionsDirectory, optionsFileName));
-        }
-
-        /// <summary>
-        /// Saves the current options to disk.
-        /// </summary>
-        public void SaveOptionsToDisk()
-        {
-            try
-            {
-                var optionsFile = GetPackageOptionsFile();
-                if (optionsFile.Directory?.Exists == false)
-                    optionsFile.Directory.Create();
-                
-                using var outputStream = optionsFile.OpenWrite();
-                JsonSerializer.Serialize(outputStream, Serialized());
-            }
-            catch (Exception ex)
-            {
-                AppTools.Log(ex);
-            }
-        }
-
-        /// <summary>
-        /// Saves the current options to disk, asynchronously.
-        /// </summary>
-        public async Task SaveOptionsToDiskAsync()
-        {
-            try
-            {
-                var optionsFile = GetPackageOptionsFile();
-                if (optionsFile.Directory?.Exists == false)
-                    optionsFile.Directory.Create();
-
-                await using var outputStream = optionsFile.OpenWrite();
-                await JsonSerializer.SerializeAsync(outputStream, Serialized());
-            }
-            catch (Exception ex)
-            {
-                AppTools.Log(ex);
-            }
-        }
-
-        /// <summary>
-        /// Loads the options from disk, asynchronously.
-        /// </summary>
-        public void LoadOptionsFromDisk()
-        {
-            try
-            {
-                var optionsFile = GetPackageOptionsFile();
-                if (!optionsFile.Exists)
-                    return;
-
-                using var inputStream = optionsFile.OpenRead();
-                var options = JsonSerializer.Deserialize<SerializableInstallationOptions_v1>(inputStream);
-                FromSerialized(options);
-            }
-            catch (Exception e)
-            {
-                AppTools.Log(e);
-            }
-        }
-
-        /// <summary>
-        /// Loads the options from disk.
-        /// </summary>
-        public async Task LoadOptionsFromDiskAsync()
-        {
-            var optionsFile = GetPackageOptionsFile();
-            try
-            {
-                if (!optionsFile.Exists)
-                    return;
-                
-
-                await using var inputStream = optionsFile.OpenRead();
-                var options = await JsonSerializer.DeserializeAsync<SerializableInstallationOptions_v1>(inputStream);
-                FromSerialized(options);
-            }
-            catch (JsonException)
-            {
-                AppTools.Log("An error occurred while parsing package " + optionsFile + ". The file will be overwritten");
-                await File.WriteAllTextAsync(optionsFile.FullName, "{}");
-            }
-            catch (Exception e)
-            {
-                AppTools.Log("Loading installation options for file " + optionsFile  + " have failed: ");
-                AppTools.Log(e);
-            }
-        }
-
-        /// <summary>
-        /// Returns a string representation of the current options.
-        /// </summary>
-        /// <returns></returns>
-        public override string ToString()
-        {
-            string customparams = CustomParameters != null ? string.Join(",", CustomParameters) : "[]";
-            return $"<InstallationOptions: SkipHashCheck={SkipHashCheck};" +
-                   $"InteractiveInstallation={InteractiveInstallation};" +
-                   $"RunAsAdministrator={RunAsAdministrator};" +
-                   $"Version={Version};" +
-                   $"Architecture={Architecture};" +
-                   $"InstallationScope={InstallationScope};" +
-                   $"InstallationScope={CustomInstallLocation};" +
-                   $"CustomParameters={customparams};" +
-                   $"RemoveDataOnUninstall={RemoveDataOnUninstall}>";
-        }
     }
 }
