+++ src/UniGetUI/PackageEngine/Classes/PackageManager.cs
@@ -1,16 +1,11 @@
-using UniGetUI.Core;
+﻿using UniGetUI.Core;
 using UniGetUI.PackageEngine.Operations;
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
-using System.Runtime.InteropServices;
 using System.Threading.Tasks;
-using Windows.Devices.Bluetooth.Advertisement;
 
 namespace UniGetUI.PackageEngine.Classes
 {
-
     /// <summary>
     /// Absract class that all managers must implement
     /// </summary>
@@ -364,197 +359,5 @@ namespace UniGetUI.PackageEngine.Classes
         /// </summary>
         /// <returns>A ManagerSource object representing the main source/index</returns>
         public abstract ManagerSource GetMainSource();
-    }
-
-    public abstract class PackageManagerWithSources : PackageManager
-    {
-        public ManagerSourceFactory SourceFactory { get; private set; }
-        public ManagerSource[] KnownSources { get; set; }
-
-        public PackageManagerWithSources() : base()
-        {
-            SourceFactory = new(this);
-        }
-
-        public virtual async Task<ManagerSource[]> GetSources()
-        {
-            try
-            {
-                ManagerSource[] sources = await GetSources_UnSafe();
-                SourceFactory.Reset();
-                foreach (ManagerSource source in sources)
-                    SourceFactory.AddSource(source);
-                return sources;
-            }
-            catch (Exception e)
-            {
-                AppTools.Log("Error finding sources for manager " + Name + ": \n" + e.ToString());
-                return new ManagerSource[0];
-            }
-        }
-
-        public abstract ManagerSource[] GetKnownSources();
-        public abstract string[] GetAddSourceParameters(ManagerSource source);
-        public abstract string[] GetRemoveSourceParameters(ManagerSource source);
-        public abstract OperationVeredict GetAddSourceOperationVeredict(ManagerSource source, int ReturnCode, string[] Output);
-        
-        public abstract OperationVeredict GetRemoveSourceOperationVeredict(ManagerSource source, int ReturnCode, string[] Output);
-        protected abstract Task<ManagerSource[]> GetSources_UnSafe();
-    }
-
-    public class ManagerStatus
-    {
-        public string Version = "";
-        public bool Found = false;
-        public string ExecutablePath = "";
-        public ManagerStatus()
-        { }
-
-    }
-    public class ManagerProperties
-    {
-        public string Name { get; set; } = "Unset";
-        public string Description { get; set; } = "Unset";
-        public string IconId { get; set; } = "Unset";
-        public string ColorIconId { get; set; } = "Unset";
-        public string ExecutableCallArgs { get; set; } = "Unset";
-        public string ExecutableFriendlyName { get; set; } = "Unset";
-        public string InstallVerb { get; set; } = "Unset";
-        public string UpdateVerb { get; set; } = "Unset";
-        public string UninstallVerb { get; set; } = "Unset";
-
-    }
-    public struct ManagerCapabilities
-    {
-        public bool CanRunAsAdmin = false;
-        public bool CanSkipIntegrityChecks = false;
-        public bool CanRunInteractively = false;
-        public bool CanRemoveDataOnUninstall = false;
-        public bool SupportsCustomVersions = false;
-        public bool SupportsCustomArchitectures = false;
-        public Architecture[] SupportedCustomArchitectures = new Architecture[0];
-        public bool SupportsCustomScopes = false;
-        public bool SupportsPreRelease = false;
-        public bool SupportsCustomLocations = false;
-        public bool SupportsCustomSources = false;
-        public ManagerSource.Capabilities Sources { get; set; }
-        public ManagerCapabilities()
-        {
-            Sources = new ManagerSource.Capabilities();
-        }
-    }
-
-    public class ManagerSource
-    {
-        public virtual string IconId { get { return Manager.Properties.IconId; } }
-        public bool IsVirtualManager = false;
-        public struct Capabilities
-        {
-            public bool KnowsUpdateDate { get; set; } = false;
-            public bool KnowsPackageCount { get; set; } = false;
-            public bool MustBeInstalledAsAdmin { get; set; } = false;
-            public Capabilities()
-            { }
-        }
-
-        public PackageManager Manager { get; }
-        public string Name { get; }
-        public Uri Url { get; private set; }
-        public int? PackageCount { get; }
-        public string UpdateDate { get; }
-
-        public ManagerSource(PackageManager manager, string name, Uri url = null, int? packageCount = 0, string updateDate = null)
-        {
-            Manager = manager;
-            Name = name;
-            Url = url;
-            if (manager.Capabilities.Sources.KnowsPackageCount)
-                PackageCount = packageCount;
-            if (manager.Capabilities.Sources.KnowsUpdateDate)
-                UpdateDate = updateDate;
-        }
-
-        public override string ToString()
-        {
-            if (Manager.Capabilities.SupportsCustomSources)
-                return Manager.Name + ": " + Name;
-            else
-                return Manager.Name;
-        }
-
-        /// <summary>
-        /// Replaces the current URL with the new one. Must be used only when a placeholder URL is used.
-        /// </summary>
-        /// <param name="newUrl"></param>
-        public void ReplaceUrl(Uri newUrl)
-        {
-            Url = newUrl;
-        }
-    }
-
-    public class ManagerSourceFactory
-    {
-        private PackageManager __manager;
-        private Dictionary<string, ManagerSource> __reference;
-        private Uri __default_uri = new Uri("https://marticliment.com/unigetui/");
-
-        public ManagerSourceFactory(PackageManager manager)
-        {
-            __reference = new();
-            __manager = manager;
-        }
-
-        public void Reset()
-        {
-            __reference.Clear();
-        }
-
-        /// <summary>
-        /// Returns the existing source for the given name, or creates a new one if it does not exist.
-        /// </summary>
-        /// <param name="name">The name of the source</param>
-        /// <returns>A valid ManagerSource</returns>
-        public ManagerSource GetSourceOrDefault(string name)
-        {
-            ManagerSource source;
-            if (__reference.TryGetValue(name, out source))
-            {
-                return source;
-            }
-
-            var new_source = new ManagerSource(__manager, name, __default_uri);
-            __reference.Add(name, new_source);
-            return new_source;
-        }
-
-        /// <summary>
-        /// Returns the existing source for the given name, or null if it does not exist.
-        /// </summary>
-        /// <param name="name"></param>
-        /// <returns></returns>
-        public ManagerSource? GetSourceIfExists(string name)
-        {
-            ManagerSource source;
-            if (__reference.TryGetValue(name, out source))
-            {
-                return source;
-            }
-            return null;
-        }
-
-        public void AddSource(ManagerSource source)
-        {
-            if(!__reference.TryAdd(source.Name, source))
-            {
-                var existing_source = __reference[source.Name];
-                if(existing_source.Url == __default_uri)
-                    existing_source.ReplaceUrl(source.Url);
-            }
-        }
-
-        public ManagerSource[] GetAvailableSources()
-        {
-            return __reference.Values.ToArray();
-        }
     }
 }
