+++ src/UniGetUI/PackageEngine/Classes/PackageTag.cs
@@ -1,0 +1,13 @@
+﻿namespace UniGetUI.PackageEngine.Classes
+{
+    public enum PackageTag
+    {
+        Default,
+        AlreadyInstalled,
+        IsUpgradable,
+        Pinned,
+        OnQueue,
+        BeingProcessed,
+        Failed
+    }
+}
