+++ src/UniGetUI/PackageEngine/Classes/SerializableIncompatiblePackage_v1.cs
@@ -1,0 +1,11 @@
+﻿namespace UniGetUI.PackageEngine.Classes
+{
+    public class SerializableIncompatiblePackage_v1
+    {
+        public string Id { get; set; } = "";
+        public string Name { get; set; } = "";
+        public string Version { get; set; } = "";
+        public string Source { get; set; } = "";
+    }
+
+}
