+++ src/UniGetUI/PackageEngine/Managers/Chocolatey.cs
@@ -6,7 +6,6 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using System.Net;
 using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 using UniGetUI.Core.Data;
