+++ src/UniGetUI/PackageEngine/Managers/Npm.cs
@@ -6,7 +6,6 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using System.Net;
 using System.Threading.Tasks;
 
 namespace UniGetUI.PackageEngine.Managers
