+++ src/UniGetUI/PackageEngine/Managers/Scoop.cs
@@ -1,5 +1,4 @@
-﻿using H.NotifyIcon.Core;
-using UniGetUI.PackageEngine.Classes;
+﻿using UniGetUI.PackageEngine.Classes;
 using UniGetUI.PackageEngine.Operations;
 using UniGetUI.Core;
 using System;
@@ -7,7 +6,6 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using System.Net;
 using System.Runtime.InteropServices;
 using System.Text.Json.Nodes;
 using System.Text.RegularExpressions;
